+++ lib/screens/splash_screen/loading_dots.dart
@@ -1,0 +1,68 @@
+import 'package:flutter/material.dart';
+
+class LoadingDots extends StatefulWidget {
+  const LoadingDots({Key? key}) : super(key: key);
+
+  @override
+  _LoadingDotsState createState() => _LoadingDotsState();
+}
+
+class _LoadingDotsState extends State<LoadingDots>
+    with TickerProviderStateMixin {
+  late List<AnimationController> _controllers;
+  late List<Animation<double>> _animations;
+
+  @override
+  void initState() {
+    super.initState();
+    _controllers = List.generate(
+      3,
+      (index) => AnimationController(
+        duration: const Duration(milliseconds: 300),
+        vsync: this,
+      ),
+    );
+    _animations = _controllers.map((controller) {
+      return Tween<double>(begin: 0, end: 6).animate(controller);
+    }).toList();
+
+    for (var i = 0; i < 3; i++) {
+      Future.delayed(Duration(milliseconds: i * 100), () {
+        _controllers[i].repeat(reverse: true);
+      });
+    }
+  }
+
+  @override
+  void dispose() {
+    for (var controller in _controllers) {
+      controller.dispose();
+    }
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: List.generate(3, (index) {
+        return AnimatedBuilder(
+          animation: _controllers[index],
+          builder: (context, child) {
+            return Container(
+              width: 8,
+              height: 8,
+              margin: const EdgeInsets.symmetric(horizontal: 4),
+              decoration: const BoxDecoration(
+                color: Color(0xFF4F46E5),
+                shape: BoxShape.circle,
+              ),
+              transform:
+                  Matrix4.translationValues(0, -_animations[index].value, 0),
+            );
+          },
+        );
+      }),
+    );
+  }
+}
